Implemented with Input handler
Actually implemented with the input handler script.
