--- Assets/Scripts/PlayerController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5f14482724a6244439f6fabac5c9e8e6
