+++ Assets/Scripts/PlayerController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 45e2c204bc92e1a418bd5ac89c4a848e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
